Refactor to the same format as other contracts
