+++ contracts/coconut-dkg/src/epoch_state/storage.rs
@@ -1,7 +1,8 @@
 // Copyright 2022 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{ContractError, ADMIN};
+use crate::error::ContractError;
+use crate::state::ADMIN;
 use coconut_dkg_common::types::EpochState;
 use cosmwasm_std::{DepsMut, MessageInfo, Response, Storage};
 use cw_storage_plus::Item;
