+++ contracts/coconut-dkg/src/error.rs
@@ -1,7 +1,7 @@
 // Copyright 2022 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-use cosmwasm_std::{Addr, StdError, VerificationError};
+use cosmwasm_std::StdError;
 use cw_controllers::AdminError;
 use thiserror::Error;
 
@@ -16,24 +16,6 @@ pub enum ContractError {
 
     #[error("Group contract invalid address '{addr}'")]
     InvalidGroup { addr: String },
-
-    #[error("No coin was sent for the deposit, you must send {denom}")]
-    NoDepositFound { denom: String },
-
-    #[error("Received multiple coin types")]
-    MultipleDenoms,
-
-    #[error("Wrong coin denomination, you must send {denom}")]
-    WrongDenom { denom: String },
-
-    #[error("Not enough funds sent for deposit. (received {received}, minimum {minimum})")]
-    InsufficientDeposit { received: u128, minimum: u128 },
-
-    #[error("Failed to perform ed25519 signature verification - {0}. This dealer will be temporarily blacklisted now.")]
-    Ed25519VerificationError(#[from] VerificationError),
-
-    #[error("Provided ed25519 signature did not verify correctly. This dealer will be temporarily blacklisted now.")]
-    InvalidEd25519Signature,
 
     #[error("This potential dealer is not in the coconut signer group")]
     Unauthorized,
@@ -51,11 +33,6 @@ pub enum ContractError {
         current_state: String,
         expected_state: String,
     },
-
-    // we should never ever see this error (famous last words in programming), therefore, I'd want to
-    // explicitly declare it so that when we ultimate do see it, it's gonna be more informative over "normal" panic
-    #[error("Somehow our validated address {address} is not using correct bech32 encoding")]
-    InvalidValidatedAddress { address: Addr },
 
     #[error("This sender is not a dealer for the current epoch")]
     NotADealer,
